Fix: 카드 뷰 분리 및 셀 재 생성#21
