+++ MUMENT/MUMENT/Sources/Components/MumentCardBySongView.swift
@@ -1,0 +1,155 @@
+//
+//  MumentCardBySongView.swift
+//  MUMENT
+//
+//  Created by 김지민 on 2022/07/15.
+//
+
+import UIKit
+import SnapKit
+import Then
+
+class MumentCardBySongView: UIView {
+    
+    // MARK: - Properties
+    lazy var writerInfoStackView = UIStackView(arrangedSubviews: [profileImage, writerNameLabel]).then{
+        $0.axis = .horizontal
+        $0.spacing = 7
+    }
+    let profileImage = UIImageView().then{
+        $0.makeRounded(cornerRadius: 12.5)
+    }
+    let writerNameLabel = UILabel().then{
+        $0.textColor = .mBlack2
+        $0.font = .mumentC1R12
+        $0.sizeToFit()
+    }
+    
+    private let heartButton = UIButton().then{
+        var configuration = UIButton.Configuration.plain()
+            configuration.imagePadding = 5
+            configuration.buttonSize = .small
+        $0.configuration = configuration
+    }
+    
+    let attributes: [NSAttributedString.Key: Any] = [
+        .font: UIFont.mumentC1R12,
+        .foregroundColor: UIColor.mGray1
+    ]
+    
+    var heartButtonText: String = "" {
+        didSet{
+            heartButton.setAttributedTitle(NSAttributedString(string: heartButtonText,attributes: attributes), for: .normal)
+        }
+    }
+    
+    let separatorView = UIView().then{
+        $0.backgroundColor = .mGray4
+    }
+    
+    ///data에 있는 것 만큼 DefaultTagView()하고 stack view에 추가
+    let tagStackView = UIStackView()
+    let contentsLabel = UILabel().then{
+        $0.textColor = .mBlack2
+        $0.lineBreakMode = .byTruncatingTail
+        $0.numberOfLines = 3
+        $0.font = UIFont(name: "NotoSans-Medium", size: 13.0)
+    }
+    let createdAtLabel = UILabel().then{
+        $0.textColor = .mGray2
+        $0.font = .mumentC1R12
+    }
+    
+    // MARK: - Initialization
+//    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+//        super.init(style: style, reuseIdentifier: reuseIdentifier)
+//        setUI()
+//        setLayout()
+//        selectionStyle = .none
+//    }
+//
+//    @available(*, unavailable)
+//    required init?(coder: NSCoder) {
+//        fatalError("init(coder:) has not been implemented")
+//    }
+    
+    override init(frame: CGRect) {
+        super.init(frame: .zero)
+//        setDefaultLayout()
+        setUI()
+        setLayout()
+//        sleselectionStyle = .none
+    }
+    
+    required init(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)!
+        setLayout()
+    }
+    
+    //MARK: - Functions
+    func setData(_ cellData: MumentCardBySongModel){
+        profileImage.image = cellData.profileImage
+        writerNameLabel.text = cellData.writerName
+        contentsLabel.text = cellData.contents
+        createdAtLabel.text = cellData.createdAt
+        heartButton.setImage(cellData.heartImage, for: .normal)
+        heartButtonText = "\(cellData.heartCount)"
+    }
+}
+
+// MARK: - UI
+extension MumentCardBySongView {
+    
+    func setUI(){
+        self.backgroundColor = .mWhite
+        self.makeRounded(cornerRadius: 11)
+        self.addShadow(offset: CGSize(width: 0, height: -2),opacity: 0.2,radius: 5.0)
+//        contentView.frame = contentView.frame.inset(by: UIEdgeInsets(top: 0, left: 20, bottom: 0, right: 0))
+
+    }
+    
+    func setLayout() {
+        self.addSubviews([writerInfoStackView,heartButton,separatorView,tagStackView,contentsLabel,createdAtLabel])
+
+        writerInfoStackView.snp.makeConstraints {
+            $0.left.equalTo(self.safeAreaLayoutGuide).offset(13)
+            $0.top.equalTo(self.safeAreaLayoutGuide).offset(11)
+        }
+        
+        heartButton.snp.makeConstraints {
+            $0.right.equalTo(self.safeAreaLayoutGuide).inset(5)
+            $0.top.equalTo(self.safeAreaLayoutGuide).offset(9)
+        }
+        
+        separatorView.snp.makeConstraints{
+            $0.left.equalTo(self.safeAreaLayoutGuide).offset(13)
+            $0.right.equalTo(self.safeAreaLayoutGuide).inset(13)
+            $0.top.equalTo(writerInfoStackView.snp.bottom).offset(11)
+            $0.height.equalTo(1)
+        }
+        
+        tagStackView.snp.makeConstraints{
+            $0.left.equalTo(self.safeAreaLayoutGuide).offset(13)
+            $0.right.equalTo(self.safeAreaLayoutGuide).inset(13)
+            $0.top.equalTo(separatorView.snp.bottom).offset(11)
+        }
+        
+        contentsLabel.snp.makeConstraints{
+            $0.left.equalTo(self.safeAreaLayoutGuide).offset(13)
+            $0.right.equalTo(self.safeAreaLayoutGuide).inset(13)
+            $0.top.equalTo(tagStackView.snp.bottom).offset(10)
+        }
+        
+        createdAtLabel.snp.makeConstraints{
+            $0.left.equalTo(self.safeAreaLayoutGuide).offset(13)
+            $0.top.equalTo(contentsLabel.snp.bottom).offset(12)
+            $0.bottom.equalTo(self.safeAreaLayoutGuide).inset(16)
+        }
+        
+
+        profileImage.snp.makeConstraints{
+            $0.height.width.equalTo(25)
+        }
+    }
+}
+
