+++ MUMENT/MUMENT/Sources/Scenes/Home/SongDetail/MyMumentSectionHeader.swift
@@ -18,42 +18,26 @@ class MyMumentSectionHeader: UITableViewHeaderFooterView {
         $0.font = .mumentB1B15
     }
     
-//    let attributes: [NSAttributedString.Key: Any] =
-    
-//    var historyButtonText: String = "" {
-//        didSet{
-//            historyButton.setAttributedTitle(NSAttributedString(string: historyButtonText,attributes: attributes), for: .normal)
-//        }
-//    }
-    
     private let historyButton = UIButton().then{
-//        $0.setImage(UIImage(named: "rightArrow"), for: .normal)
         $0.configuration = .plain()
         $0.configuration?.imagePlacement = .trailing
-//        $0.setTitle("나의 히스토리 보기 >", for: .normal)
         $0.setAttributedTitle(NSAttributedString(string: "나의 히스토리 보기 >", attributes: [
             .font: UIFont.mumentC1R12,
             .foregroundColor: UIColor.mPurple1
         ]), for: .normal)
     }
     
-
+    
     // MARK: - Initialization
     override init(reuseIdentifier: String?) {
-            super.init(reuseIdentifier: reuseIdentifier)
-            setLayout()
-//            configureLayout()
-        }
-        
-        required init?(coder: NSCoder) {
-            super.init(coder: coder)
-            setLayout()
-        }
+        super.init(reuseIdentifier: reuseIdentifier)
+        setLayout()
+    }
     
-//    //MARK: - Functions
-//    func setTitle(_ title:String){
-//        titleLabel.text = title
-//    }
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        setLayout()
+    }
 }
 
 // MARK: - UI
@@ -63,15 +47,13 @@ extension MyMumentSectionHeader {
         contentView.addSubviews([titleLabel,historyButton])
         
         titleLabel.snp.makeConstraints{
-            $0.leading.equalTo(self.safeAreaLayoutGuide)
+            $0.leading.equalTo(self.safeAreaLayoutGuide).offset(20)
             $0.top.equalTo(self.safeAreaLayoutGuide).offset(15)
-//            $0.height.width.equalTo(48)
         }
         
         historyButton.snp.makeConstraints{
             $0.top.equalTo(self.safeAreaLayoutGuide).offset(10)
-            $0.trailing.equalTo(self.safeAreaLayoutGuide)
-//            $0.centerX.centerY.equalTo(self.safeAreaLayoutGuide)
+            $0.trailing.equalTo(self.safeAreaLayoutGuide).inset(20)
         }
     }
 }
