+++ MUMENT/MUMENT/Sources/Scenes/Home/SongDetail/SongDetailTVC.swift
@@ -26,9 +26,6 @@ class SongDetailTVC: UITableViewCell {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
-    // MARK: - Functions
-    
 }
 
 // MARK: - UI
